Correcting the code to one style, UniversityContext.OnModelCreating() - search in assembly, Add development appsettings

+++ WpfUniversity/Command/Courses/DeleteCourseCommand.cs
@@ -1,51 +1,46 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-using UniversityDataLayer.Entities;
 using WpfUniversity.Services;
 using WpfUniversity.Services.Courses;
 using WpfUniversity.ViewModels.Courses;
 using WpfUniversity.ViewModels.Dialogs;
 
-namespace WpfUniversity.Command.Courses
+namespace WpfUniversity.Command.Courses;
+
+public class DeleteCourseCommand : AsyncCommandBase
 {
-    public class DeleteCourseCommand : AsyncCommandBase
+    private readonly CourseViewModel _courseViewModel;
+    private readonly CourseService _courseService;
+    private readonly ModalNavigationService _modalNavigationService;
+
+    public DeleteCourseCommand(CourseViewModel courseViewModel, CourseService courseService, ModalNavigationService modalNavigationService)
     {
-        private readonly CourseViewModel _courseViewModel;
-        private readonly CourseService _courseService;
-        private readonly ModalNavigationService _modalNavigationService;
+        _courseViewModel = courseViewModel;
+        _courseService = courseService;
+        _modalNavigationService = modalNavigationService;
+    }
 
-        public DeleteCourseCommand(CourseViewModel courseViewModel, CourseService courseService, ModalNavigationService modalNavigationService)
+    public override async Task ExecuteAsync(object parameter)
+    {
+        // todo: Messange.Show in ViewModel, is it a good idea?
+        /*
+        YesNoDialogViewModel dialog = new YesNoDialogViewModel(_modalNavigationService);
+        dialog.Message = "";        
+        dialog.ShowDialogCommand.Execute(null);
+        if (!dialog.UserChoice) return; 
+        */
+
+        _courseViewModel.ErrorMessage = null;
+        var course = _courseViewModel.CourseTreeViewModel.SelectedCourse;
+
+        try
         {
-            _courseViewModel = courseViewModel;
-            _courseService = courseService;
-            _modalNavigationService = modalNavigationService;
+            await _courseService.Delete(course);
+        }
+        catch (Exception e)
+        {
+            _courseViewModel.ErrorMessage = e.Message;
         }
 
-        public override async Task ExecuteAsync(object parameter)
-        {
-            // todo: This show method yes/no dialog doesn't work
-            /*
-            YesNoDialogViewModel dialog = new YesNoDialogViewModel(_modalNavigationService);
-            dialog.Message = "Do you want to proceed?";
-
-            _modalNavigationService.CurrentViewModel = dialog;
-            */
-
-            _courseViewModel.ErrorMessage = null;
-            var course = _courseViewModel.CourseTreeViewModel.SelectedCourse;
-
-            try
-            {
-                await _courseService.Delete(course);
-            }
-            catch (Exception e)
-            {
-                _courseViewModel.ErrorMessage = e.Message;
-            }
-
-        }
     }
 }
